+++ app/model/services/ConsultationManager.scala
@@ -97,7 +97,7 @@ class ConsultationManager {
   }
 
 
-  def getOpenGovComments(consultationId:Int,articleId:Int, maxCommentId:Option[Long]): List[Comment] = {
+  def getOpenGovComments(consultationId:Long,articleId:Long, maxCommentId:Option[Long]): List[Comment] = {
 
     val commentsRepository = new CommentsRepository()
     commentsRepository.getComments(consultationId,articleId,None,CommentSource.OpenGov,maxCommentId,commentsPageSize )
